+++ src/main/java/basemod/interfaces/AddCustomModeModsSubscriber.java
@@ -4,6 +4,6 @@ import com.megacrit.cardcrawl.screens.custom.CustomMod;
 
 import java.util.List;
 
-public interface EditCustomModeModsSubscriber extends ISubscriber {
+public interface AddCustomModeModsSubscriber extends ISubscriber {
     void receiveCustomModeMods(List<CustomMod> modList);
 }